Add variable: switch-window-input-style
Switch-window use read-event to read users input by default, it will
be conflicted with EXWM (Emacs X window manager), So I add a new
input style: 'minibuffer to handle this issue.

The origin input style is named as 'default, the intuitive different
is that: 'minibuffer style need to "confirm" your input with Enter.